User management APIs

+++ app/controllers/sessions_controller.rb
@@ -2,7 +2,7 @@ class SessionsController < ApplicationController
   skip_before_action :verify_authenticity_token, only: :create
 
   def new
-    (redirect_to user_url) && return if current_user
+    return (redirect_to user_url) if current_user
 
     login_name = session.delete(:unauthorized_user)
     if login_name.present?
